[webapp-test-bridge] Address review comments
  1. withCredentials → withSecrets: Added withSecrets.githubActionsDispatch() to withSecrets.groovy that fetches GITHUB_TOKEN from GSM and exposes it as an env var. runGithubAction.groovy now
  wraps both dispatch and wait in this block.
  2. sh → exec for shell escaping: Both curl calls now use exec([...]) and exec.outputOf([...]), and gh run watch uses exec([...]) — all arguments are passed as list elements, no manual shell
  escaping needed.
  3. def call(Map args) that dispatch+waits: Added so callers just call runGithubAction(...) rather than the two separate methods.
  4. Renamed ghActions.groovy → runGithubAction.groovy: Old file deleted, webapp-test.groovy updated to call runGithubAction(...).
  5. Added runGithubAction.txt documentation file.
  6. make-allcheck.groovy now passes USE_GITHUB_BRIDGE=true for nightly runs.
