+++ vars/withSecrets.txt
@@ -11,3 +11,5 @@ slack.  Here are all the cases:
 * slackAlertlibOnly
 * githubAlertlibOnly
 * slackAndStackdriverAlertlibOnly
+* githubActionsDispatch -- exposes GITHUB_TOKEN env var for GitHub API
+  calls and the gh CLI (used by runGithubAction)
